+++ src/python/deepgnn/graph_engine/data/citeseer.py
@@ -38,14 +38,14 @@ class CiteseerFull(Dataset):
     ):
         """Initialize dataset."""
         super().__init__(
-            input_location=f"{str(Path(__file__).parent)}",
+            f"/home/user/DeepGNNMain/src/python/deepgnn/graph_engine/data" if url == "local" else url,
             name="citeseer",
             num_nodes=3312,
             feature_dim=3703,
             num_classes=6,
             train_node_ratio=train_node_ratio,
             random_selection=random_selection,
-            output_dir=f"{str(Path(__file__).parent)}",
+            output_dir=f"{str(Path(__file__).parent)}" if output_dir is None else output_dir,
         )
 
     def _load_raw_graph(
@@ -56,8 +56,8 @@ class CiteseerFull(Dataset):
         DefaultDict[int, Set[int]],
         DefaultDict[int, Set[int]],
     ]:
-        node_file = os.path.join(data_dir, "citeseer.content")
-        edge_file = os.path.join(data_dir, "citeseer.cites")
+        node_file = os.path.join(data_dir, "citeseer-doc-classification", "citeseer.content")
+        edge_file = os.path.join(data_dir, "citeseer-doc-classification", "citeseer.cites")
         paper_nodeid = {}
         labelid = {}  # type: ignore
         nodes = {}